+++ r8s/r8s_group/parent_shaperule.py
@@ -60,7 +60,7 @@ def add(parent_id, cloud, action, condition, field, value):
 
 
 @shape_rule.command(cls=ViewCommand, name='update')
-@click.option('--rule_id', '-rid', type=str,
+@click.option('--rule_id', '-rid', type=str, required=True,
               help='Shape rule id to update.')
 @click.option('--parent_id', '-pid', type=str,
               help='Parent id to update shape rule in.')
@@ -111,10 +111,13 @@ def delete(rule_id, parent_id):
 @shape_rule.command(cls=ViewCommand, name='dry_run')
 @click.option('--parent_id', '-pid', type=str, required=True,
               help='Parent id to perform dry run on shape rules')
+@click.option('--cloud', '-c', type=click.Choice(AVAILABLE_CLOUDS),
+              required=True, help='Cloud to perform dry run.')
 @cli_response()
-def dry_run(parent_id):
+def dry_run(parent_id, cloud):
     """
-    Describes shapes that satisfy all of the specified Parent rules.
+    Describes shapes that satisfy all of the specified Parent rules for cloud.
     """
     from r8s_service.initializer import init_configuration
-    return init_configuration().shape_rule_dry_run_get(parent_id=parent_id)
+    return init_configuration().shape_rule_dry_run_get(
+        parent_id=parent_id, cloud=cloud)