+++ src/services/shape_rule_filter_service.py
@@ -1,6 +1,7 @@
 import re
 from typing import List
 
+from commons.constants import CLOUD_ATTR
 from commons.log_helper import get_logger
 from models.parent_attributes import ParentMeta
 from models.shape import Shape
@@ -35,6 +36,9 @@ class ShapeRulesFilterService:
                                    parent_meta: ParentMeta,
                                    instances_data: List[Shape]):
         shape_rules = parent_meta.shape_rules
+
+        shape_rules = [rule for rule in shape_rules
+                       if rule.get(CLOUD_ATTR) == cloud]
 
         allow_filters = self.filter_by_action(
             shape_rules=shape_rules,