+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.13.0] - 2023-08-10
+* update `r8s parent shape_rule dry_run` command: add required `--cloud` parameter.
+* update `r8s parent shape_rule update` command: make `--rule_id` parameter required.
+
 ## [2.12.5] - 2023-08-10
 * fix invalid Tenant `pid` attribute extraction on job submit